requirements.txt
flask==2.3.2
openai==1.3.5
pandas==2.0.3
python-dotenv==1.0.0